+++ src/calc/calcContrastFromPreparedColors.ts
@@ -1,4 +1,5 @@
 import { ColorSpace, type ContrastModel, type PreparedColor } from "../types";
+import { ASSERT_EXAUSTED } from "../utils/assert";
 import { calcApcaP3 } from "./calcApcaP3";
 import { calcApcaSrgb } from "./calcApcaSrgb";
 import { calcWcag } from "./calcWcag";
@@ -9,19 +10,21 @@ export function calcContrastFromPreparedColors(
   contrastModel: ContrastModel,
   colorSpace: ColorSpace
 ) {
-  switch (contrastModel) {
-    case "apca": {
-      if (colorSpace === "p3") {
-        return calcApcaP3(fgColor, bgColor);
-      } else {
-        return calcApcaSrgb(fgColor, bgColor);
-      }
-    }
-    case "wcag":
-      return calcWcag(fgColor, bgColor);
-    default:
-      throw new Error(
-        'Invalid contrast model. Suported models: "apca", "wcag"'
-      );
+  // apca
+  if (contrastModel === "apca") {
+    if (colorSpace === "p3") return calcApcaP3(fgColor, bgColor);
+    if (colorSpace === "rgb") return calcApcaSrgb(fgColor, bgColor);
+    return ASSERT_EXAUSTED(colorSpace, "unknown colorSpace");
   }
+
+  // wcag
+  // 💬 2024-06-17 rvion:
+  // | is it the same scale as the one used in the APCA ?
+  if (contrastModel === "wcag") return calcWcag(fgColor, bgColor);
+
+  //
+  return ASSERT_EXAUSTED(
+    contrastModel,
+    'Invalid contrast model. Suported models: "apca", "wcag"'
+  );
 }
