+++ src/light/lightnessAndPatch.ts
@@ -1,5 +1,5 @@
-import { type ContrastConfig_WTF } from "../contrastConfig";
-import { convertToOklch_orThrow } from "../culoriUtils";
+import { type ContrastConfig_WTF } from "../contrast/contrastConfig";
+import { convertToOklch_orThrow } from "../utils/culoriUtils";
 import { log } from "../utils/log";
 
 export function lightnessAndPatch(contrastConfig: ContrastConfig_WTF) {
